criado controller para hora trabalhada

+++ src/main/scala/br/com/fabioaraujo/calcula/produto/dto/HoraTrabalhada.scala
@@ -7,6 +7,6 @@ import scala.beans.BeanProperty
 @Entity
 class HoraTrabalhada {
   @BeanProperty
-  @Id val id: Int = 1
+  @Id val id: Long = 1
   @BeanProperty var custo: Double = 10.0
 }
